NetCDFReader code to be moved into ExperimentalData class
